main: first pass entry points and resulting changes.
Quite a lot of nastiness caused by the switch to spawning dispatches when due
(rather than in advance) and trying to cope with edge cases which might be
caused by code and data diverging.

In the end, we use another registerUtility pattern to store and then find
the configured dispatch data. This necessitated much jiggery pokery.